+++ plugins/source/aws/resources/services/elasticache/clusters.go
@@ -391,7 +391,6 @@ func Clusters() *schema.Table {
 
 func fetchElasticacheClusters(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- interface{}) error {
 	awsProviderClient := meta.(*client.Client)
-
 	svc := awsProviderClient.Services().ElastiCache
 
 	var describeCacheClustersInput elasticache.DescribeCacheClustersInput